requirements.txt
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dash==1.21.0
dash-bootstrap-components==0.12.2
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-renderer==1.8.1
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
idna==3.2
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
nltk==3.6.2
numpy==1.21.0
oauthlib==3.1.1
pandas==1.3.0
pandas-datareader==0.10.0
Pillow==8.3.1
plotly==5.1.0
protobuf==3.17.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
retrying==1.3.3
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
tenacity==8.0.1
threadpoolctl==2.2.0
tqdm==4.61.2
tweepy==3.10.0
urllib3==1.26.6
Werkzeug==2.0.1
xlrd==1.2.0